Refactor GetLocalIP to iptool.GetLocalIP
