+++ pkg/service/metadata/drone/utils.go
@@ -6,25 +6,11 @@ package drone
 
 import (
 	"fmt"
-	"net"
 	"strings"
+
+	"openpitrix.io/openpitrix/pkg/utils/iptool"
 )
 
 func MakeDroneId(suffix string) string {
-	return fmt.Sprintf("drone@%s/%s", getLocalIP(), strings.TrimSpace(suffix))
-}
-
-func getLocalIP() string {
-	addrs, err := net.InterfaceAddrs()
-	if err != nil {
-		return "127.0.0.1"
-	}
-	for _, address := range addrs {
-		if ipnet, ok := address.(*net.IPNet); ok && !ipnet.IP.IsLoopback() {
-			if ipnet.IP.To4() != nil {
-				return ipnet.IP.String()
-			}
-		}
-	}
-	return "127.0.0.1"
+	return fmt.Sprintf("drone@%s/%s", iptool.GetLocalIP(), strings.TrimSpace(suffix))
 }
